Cleaning up test code
