+++ tests/CMakeLists.txt
@@ -3,6 +3,8 @@ SET(SRCROOT "${PROJECT_SOURCE_DIR}/tests")
 SET(SRC
     "${SRCROOT}/main.cpp"
 
+    "${SRCROOT}/comprehension.cpp"
+
     "${SRCROOT}/component.cpp"
     "${SRCROOT}/entity.cpp"
     "${SRCROOT}/speed.cpp"
